src/todos/utils/boolean-transformer.ts
/**
 * Copied from Stackoverflow https://stackoverflow.com/questions/59046629/boolean-parameter-in-request-body-is-always-true-in-nestjs-api
 * Credit to David Kerr https://stackoverflow.com/users/3951121/david-kerr
 * There would be great if there is a built in Transformer which would convert string to to boolean
 * In the customized transformer ToBoolean, it took any string other than ['true', 'on', 'yes', '1'] or ['false', 'off', 'no', '0'] to null
 */

import { Transform } from 'class-transformer';

export const ToBoolean = () => {
  const toPlain = Transform(
    ({ value }) => {
      return value;
    },
    {
      toPlainOnly: true,
    },
  );
  const toClass = (target: any, key: string) => {
    return Transform(
      ({ obj }) => {
        return valueToBoolean(obj[key]);
      },
      {
        toClassOnly: true,
      },
    )(target, key);
  };
  return function (target: any, key: string) {
    toPlain(target, key);
    toClass(target, key);
  };
};

const valueToBoolean = (value: any) => {
  if (value === null || value === undefined) {
    return undefined;
  }
  if (typeof value === 'boolean') {
    return value;
  }
  if (['true', 'on', 'yes', '1'].includes(value.toLowerCase())) {
    return true;
  }
  if (['false', 'off', 'no', '0'].includes(value.toLowerCase())) {
    return false;
  }
  return undefined;
};
